fileActions: split this into many small files
This large file was a catch-all for quite a lot of functionality, but
we've been slowly splitting things out of it (mostly into dialogs/ and
menu.tsx).  Let's finish the job with the four things left.

  - src/dialogs/delete.tsx: the delete dialog
  - src/dialogs/rename.tsx: the rename dialog
  - src/dialogs/permissions.jsx: the permissions dialog
  - src/download.tsx: file downloading

The permissions dialog is the only one where porting it to TypeScript
would have been non-trivial, so I avoided it for the time being.

Make some minor adjustments to the import statements where this
functionality is used, in the dropdown menu, files-card-body, and the
sidebar.
